refactor: 컴포넌트 생성 및 로직 분리
